Updating and adding tests
