+++ graylog2-web-interface/src/views/pages/DashboardsPage.test.tsx
@@ -15,18 +15,7 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { render, screen, waitFor } from 'wrappedTestingLibrary';
-import moment from 'moment';
-import userEvent from '@testing-library/user-event';
-import { PluginStore } from 'graylog-web-plugin/plugin';
-
-import type { DashboardsStoreState } from 'views/stores/DashboardsStore';
-import { asMock, MockStore } from 'helpers/mocking';
-import Routes from 'routing/Routes';
-import useDashboards from 'views/logic/dashboards/useDashboards';
-import { simpleView } from 'views/test/ViewFixtures';
-import UserDateTimeContext from 'contexts/UserDateTimeContext';
-import { ViewManagementActions } from 'views/stores/ViewManagementStore';
+import { render, screen } from 'wrappedTestingLibrary';
 
 import DashboardsPage from './DashboardsPage';
 
@@ -34,193 +23,11 @@ jest.mock('routing/Routes', () => ({
   pluginRoute: jest.fn(),
 }));
 
-jest.mock('views/logic/dashboards/useDashboards');
-
-jest.mock('views/stores/ViewManagementStore', () => ({
-  ViewManagementActions: {
-    delete: jest.fn(),
-  },
-}));
-
-jest.mock('views/stores/DashboardsStore', () => ({
-  DashboardsActions: {
-    search: () => Promise.resolve(),
-  },
-  DashboardsStore: MockStore(),
-}));
-
-const SUT = () => (
-  <UserDateTimeContext.Provider value={{
-    formatTime: (dateTime) => dateTime.toString(),
-    toUserTimezone: (time) => moment(time),
-    userTimezone: 'Europe/Paris',
-  }}>
-    <DashboardsPage />
-  </UserDateTimeContext.Provider>
-);
-
-const noDashboards: DashboardsStoreState = {
-  list: [],
-  pagination: { count: 0, page: 1, perPage: 10, total: 0 },
-};
-
-const simpleDashboardList: DashboardsStoreState = {
-  list: [simpleView()],
-  pagination: { count: 1, page: 1, perPage: 10, total: 1 },
-};
-
-const mockDashboards = (dashboardState: DashboardsStoreState) => {
-  asMock(useDashboards).mockReturnValue(dashboardState);
-};
-
-const clickDashboardAction = async (dashboardId: string, action: string) => {
-  const actionsButton = (await screen.findAllByTestId(`dashboard-actions-dropdown-${dashboardId}`))[0];
-  userEvent.click(actionsButton);
-
-  userEvent.click((await screen.findAllByRole('menuitem', { name: action }))[0]);
-};
-
 describe('DashboardsPage', () => {
-  let oldWindowConfirm;
+  it('should render header and list', async () => {
+    render(<DashboardsPage />);
 
-  beforeEach(() => {
-    oldWindowConfirm = window.confirm;
-    window.confirm = jest.fn();
-
-    asMock(Routes.pluginRoute).mockImplementation((key: string) => {
-      switch (key) {
-        case 'DASHBOARDS_NEW':
-          return '/dashboards/new';
-        case 'DASHBOARDS_VIEWID':
-          return (id: string) => `/dashboards/${id}`;
-        default:
-          throw Error(`Invalid route: ${key}`);
-      }
-    });
-  });
-
-  afterEach(() => {
-    window.confirm = oldWindowConfirm;
-  });
-
-  it('shows placeholder text if dashboards are empty', async () => {
-    mockDashboards(noDashboards);
-
-    render(<SUT />);
-
-    await screen.findByText(/Create a new dashboard here/);
-  });
-
-  it('shows list of dashboards', async () => {
-    mockDashboards(simpleDashboardList);
-    render(<SUT />);
-
-    await screen.findByRole('link', { name: 'Foo' });
-  });
-
-  it('does not delete dashboard when user clicks cancel', async () => {
-    asMock(window.confirm).mockReturnValue(false);
-
-    render(<SUT />);
-
-    await clickDashboardAction('foo', 'Delete');
-
-    await waitFor(() => expect(window.confirm).toHaveBeenCalledWith('Are you sure you want to delete "Foo"?'));
-
-    expect(ViewManagementActions.delete).not.toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' }));
-  });
-
-  it('deletes dashboard when user confirms deletion', async () => {
-    asMock(window.confirm).mockReturnValue(true);
-
-    render(<SUT />);
-
-    await clickDashboardAction('foo', 'Delete');
-
-    await waitFor(() => expect(window.confirm).toHaveBeenCalledWith('Are you sure you want to delete "Foo"?'));
-
-    expect(ViewManagementActions.delete).toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' }));
-  });
-
-  describe('supports dashboard deletion hook', () => {
-    const deletingDashboard = jest.fn(() => Promise.resolve(true));
-
-    const plugin = {
-      exports: {
-        'views.hooks.confirmDeletingDashboard': [deletingDashboard],
-      },
-    };
-
-    beforeEach(() => {
-      PluginStore.register(plugin);
-      asMock(ViewManagementActions.delete).mockClear();
-      asMock(ViewManagementActions.delete).mockImplementation((view) => Promise.resolve(view));
-      mockDashboards(simpleDashboardList);
-    });
-
-    afterEach(() => {
-      PluginStore.unregister(plugin);
-    });
-
-    it('triggers hook when deleting dashboard', async () => {
-      render(<SUT />);
-
-      await clickDashboardAction('foo', 'Delete');
-
-      await waitFor(() => expect(ViewManagementActions.delete).toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' })));
-
-      expect(deletingDashboard).toHaveBeenCalledWith(simpleDashboardList.list[0]);
-    });
-
-    it('does not delete dashboard when hook returns false', async () => {
-      asMock(deletingDashboard).mockResolvedValue(false);
-
-      render(<SUT />);
-
-      await clickDashboardAction('foo', 'Delete');
-
-      await waitFor(() => expect(deletingDashboard).toHaveBeenCalledWith(simpleDashboardList.list[0]));
-
-      expect(ViewManagementActions.delete).not.toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' }));
-    });
-
-    it('resorts to default behavior when hook returns `null`', async () => {
-      asMock(deletingDashboard).mockReturnValue(null);
-      asMock(window.confirm).mockReturnValue(true);
-
-      render(<SUT />);
-
-      await clickDashboardAction('foo', 'Delete');
-
-      await waitFor(() => expect(deletingDashboard).toHaveBeenCalledWith(simpleDashboardList.list[0]));
-
-      expect(window.confirm).toHaveBeenCalledWith('Are you sure you want to delete "Foo"?');
-
-      expect(ViewManagementActions.delete).toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' }));
-    });
-
-    it('resorts to default behavior when hook throws error', async () => {
-      const error = Error('Boom!');
-      asMock(deletingDashboard).mockImplementation(() => { throw error; });
-      asMock(window.confirm).mockReturnValue(true);
-
-      render(<SUT />);
-
-      /* eslint-disable no-console */
-      const oldConsoleTrace = console.trace;
-      console.trace = jest.fn();
-
-      await clickDashboardAction('foo', 'Delete');
-
-      await waitFor(() => expect(console.trace).toHaveBeenCalledWith('Exception occurred in deletion confirmation hook: ', error));
-      console.trace = oldConsoleTrace;
-      /* eslint-enable no-console */
-
-      await waitFor(() => expect(deletingDashboard).toHaveBeenCalledWith(simpleDashboardList.list[0]));
-
-      expect(window.confirm).toHaveBeenCalledWith('Are you sure you want to delete "Foo"?');
-
-      expect(ViewManagementActions.delete).toHaveBeenCalledWith(expect.objectContaining({ id: 'foo' }));
-    });
+    await screen.findByRole('heading', { name: /dashboards/i });
+    await screen.findByText('No dashboards have been created yet.');
   });
 });
